AXON-1246: Add mentions to createIssue page and move common logic into parent class
